Revert "Move most KnownRuntimeEffects into the sksl_rt_shader module."
This reverts commit 4db44b0c04d56caeb9f46075b889037fe7e66265.

Reason for revert: possibly breaking Chrome roll

Original change's description:
> Move most KnownRuntimeEffects into the sksl_rt_shader module.
>
> The remaining KnownRuntimeEffects are assembled programmatically,
> or are so simple that they wouldn't benefit from existing in a
> module.
>
> Bug: b/328051548
> Reviewed-on: https://skia-review.googlesource.com/c/skia/+/883098
> Commit-Queue: John Stiles <johnstiles@google.com>
> Commit-Queue: Robert Phillips <robertphillips@google.com>
> Auto-Submit: John Stiles <johnstiles@google.com>

Bug: b/328051548
No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/883878
Commit-Queue: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Bot-Commit: Rubber Stamper <rubber-stamper@appspot.gserviceaccount.com>
Auto-Submit: John Stiles <johnstiles@google.com>

+++ src/effects/imagefilters/SkMorphologyImageFilter.cpp
@@ -25,6 +25,7 @@
 #include "src/core/SkWriteBuffer.h"
 
 #include <algorithm>
+#include <cstdint>
 #include <optional>
 #include <utility>
 
@@ -127,7 +128,7 @@ sk_sp<SkImageFilter> make_morphology(MorphType type,
 // room. The other tradeoff is that for R > kMaxLinearRadius, the sparse morphology kernel only
 // requires 2 samples to double the accumulated kernel size, but at the cost of another render
 // target.
-static constexpr int kMaxLinearRadius = 14; // KEEP IN SYNC WITH CONSTANT IN `sksl_rt_shader.sksl`
+static constexpr int kMaxLinearRadius = 14; // KEEP IN SYNC W/ SkKnownRuntimeEffects.cpp VERSION
 sk_sp<SkShader> make_linear_morphology(sk_sp<SkShader> input,
                                        MorphType type,
                                        MorphDirection direction,
@@ -141,7 +142,7 @@ sk_sp<SkShader> make_linear_morphology(sk_sp<SkShader> input,
     builder.child("child") = std::move(input);
     builder.uniform("offset") = direction == MorphDirection::kX ? SkV2{1.f, 0.f} : SkV2{0.f, 1.f};
     builder.uniform("flip") = (type == MorphType::kDilate) ? 1.f : -1.f;
-    builder.uniform("radius") = (float)radius;
+    builder.uniform("radius") = (int32_t) radius;
 
     return builder.makeShader();
 }
